Extract more logic into hooks
